requirements.txt
alabaster==0.7.10
apns2==0.3.0
asn1crypto==0.23.0
astroid==1.5.3
Babel==2.5.1
bleach==2.1.1
certifi==2017.7.27.1
cffi==1.11.1
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.0.3
cryptography==2.0.3
decorator==4.1.2
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-braces==1.11.0
django-cleanup==1.0.1
django-fcm==0.1.1
django-filter==1.0.4
django-filters==0.2.1
django-push-notifications==1.5.0
django-rest-auth==0.9.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.3.2
docutils==0.14
drfdocs==0.0.11
entrypoints==0.2.3
flake8==2.5.1
h2==2.6.2
hpack==3.0.0
html5lib==1.0b10
hyper==0.7.0
hyperframe==3.2.0
idna==2.6
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.3.1
livereload==2.5.1
Markdown==2.6.9
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.7.4
mkdocs==0.15.3
mkdocs-bootstrap==0.1.1
mkdocs-bootswatch==0.4.0
nbconvert==5.3.1
nbformat==4.4.0
numpydoc==0.7.0
oauthlib==2.0.4
olefile==0.44
pandocfilters==1.4.2
parso==0.1.0
pep8==1.7.0
pickleshare==0.7.4
Pillow==4.2.1
pluggy==0.5.2
prompt-toolkit==1.0.15
psutil==5.3.1
py==1.4.34
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.7.4
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
requests-oauthlib==0.8.0
rope==0.10.7
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
Sphinx==1.6.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.3
testpath==0.3.1
tornado==4.5.2
tox==2.8.2
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.10.11
